gpio_input: add accidental double press fix to it
Replicating the gpio_matrix higher level fix to the
simple input driver. This should keep the twitchy
power button at bay (pressing the power button
once resulted in screen on/off/on as the driver
put two events subsequently instead of on.)
This also fixes the annoying 'long power menu +
unwanted screen off right after it' bug too,
that was produced by this doubleing as well.

Change-Id: Iafa8922972c02d8b9e7c2572a401389b83865c7c